requirements-celery.txt
celery==5.0.0
python-decouple==3.3
redis==3.5.3
